refactor: 컴포넌트 명 index로 변경해서 파일 경로 단축 및 적용
