+++ src/app/(home)/_components/LandingPage.tsx
@@ -8,8 +8,8 @@ import Person from '@/assets/image/hero_person.png';
 import DescImageOne from '@/assets/image/desc_image_1.png';
 import DescImageTwo from '@/assets/image/desc_image_2.png';
 import Image from 'next/image';
-import Button from '@/components/Button/Button';
-import Footer from '@/components/Footer/Footer';
+import Button from '@/components/Button';
+import Footer from '@/components/Footer';
 import BlackArrow from '@/assets/icon/arrow-black.svg';
 import {loginKaKao} from '@/api/user';
