+++ src/app/rehearsal/_components/RehearsalHeader/index.tsx
