+++ src/app/rehearsal/_components/Timer/index.tsx
