+++ src/app/rehearsal/setting/_components/SoundBar/index.tsx
