+++ src/app/users/oauth/kakao/page.tsx
@@ -1,9 +1,10 @@
 'use client';
+
 import {useEffect} from 'react';
 import {getKakaoProfile} from '@/api/user';
 import {useUserStore} from '@/stores/userStore';
 import {useRouter} from 'next/navigation';
-import Loader from '@/components/Loader/Loader';
+import Loader from '@/components/Loader';
 
 const KakaoRedirectPage = () => {
   let code = '';
